+++ client/views/posts/post_item.js
@@ -3,12 +3,11 @@ Template.post_item.created = function () {
 };
 
 Template.post_item.helpers({
-  post: function(){
-    // note: when the data context is set by the router, it will be "this.post". When set by a parent template it'll be "this"
-    return this.post || this;
-  },
   postLink: function(){
     return !!this.url ? getOutgoingUrl(this.url) : "/posts/"+this._id;
+  },
+  sourceLink: function(){
+    return !!this.url ? this.url : "/posts/"+this._id;
   },
   postTarget: function() {
     return !!this.url ? '_blank' : '';
@@ -40,18 +39,13 @@ Template.post_item.helpers({
   short_score: function(){
     return Math.floor(this.score*1000)/1000;
   },
-  body_formatted: function(){
-    var converter = new Markdown.Converter();
-    var html_body=converter.makeHtml(this.body);
-    return html_body.autoLink();
-  },
   ago: function(){
     // if post is approved show submission time, else show creation time. 
-    time = this.status == STATUS_APPROVED ? this.submitted : this.createdAt;
+    time = this.status == STATUS_APPROVED ? this.postedAt : this.createdAt;
     return moment(time).fromNow();
   },
   timestamp: function(){
-    time = this.status == STATUS_APPROVED ? this.submitted : this.createdAt;
+    time = this.status == STATUS_APPROVED ? this.postedAt : this.createdAt;
     return moment(time).format("MMMM Do, h:mm:ss a");
   },
   voted: function(){
@@ -74,45 +68,15 @@ Template.post_item.helpers({
     return this.comments == 1 ? i18n.t('comment') : i18n.t('comments');
   },
   pointsUnitDisplayText: function(){
-    return this.votes == 1 ? i18n.t('point') : i18n.t('points');
+    return this.upvotes == 1 ? i18n.t('point') : i18n.t('points');
   },
   isApproved: function(){
     return this.status == STATUS_APPROVED;
+  },
+  viaTwitter: function () {
+    return !!getSetting('twitterAccount') ? 'via='+getSetting('twitterAccount') : '';
   }
 });
-
-var recalculatePosition = function ($object, pArray) {
-  // delta is the difference between the last two positions in the array
-  var delta = pArray[pArray.length-2] - pArray[pArray.length-1];
-
-  // if new position is different from previous position
-  if(delta != 0){
-    // send object back to previous position
-    $object.removeClass('animate').css("top", delta + "px");
-    // then wait a little and animate it to new one
-    setTimeout(function() { 
-      $object.addClass('animate').css("top", "0px")
-    }, 1);  
-  }
-}
-
-Template.post_item.rendered = function(){
-  // var instance = this,
-  //     $instance = $(instance.firstNode.nextSibling),
-  //     top = $instance.position().top;
-
-  // // if this is the first render, initialize array, else push current position
-  // if(typeof instance.pArray === 'undefined'){
-  //   instance.pArray = [top]
-  // }else{
-  //   instance.pArray.push(top);
-  // }
-
-  // // if this is *not* the first render, recalculate positions
-  // if(instance.pArray.length>1)
-  //   recalculatePosition($instance, instance.pArray);
-
-};
 
 Template.post_item.events({
   'click .upvote-link': function(e, instance){
@@ -134,7 +98,6 @@ Template.post_item.events({
     $(".share-options").not($share).addClass("hidden");
     $this.toggleClass("active");
     $share.toggleClass("hidden");
-    $share.find('.share-replace').sharrre(SharrreOptions);
   },
   'click .approve-link': function(e, instance){
     Meteor.call('approvePost', this);