+++ client/views/posts/post_page.js
@@ -1,17 +1,3 @@
-Template.post_page.helpers({
-  post: function () {
-    return Posts.findOne(this.postId);
-  },
-  body_formatted: function(){
-    var converter = new Markdown.Converter();
-    var html_body=converter.makeHtml(this.body);
-    return html_body.autoLink();
-  },
-  canComment: function(){
-    return canComment(Meteor.user());
-  }
-}); 
-
 Template.post_page.rendered = function(){
   if((scrollToCommentId=Session.get('scrollToCommentId')) && !this.rendered && $('#'+scrollToCommentId).exists()){
     scrollPageTo('#'+scrollToCommentId);
@@ -19,5 +5,5 @@ Template.post_page.rendered = function(){
     this.rendered=true;
   }
   if(this.data) // XXX
-    document.title = this.data.headline;
+    document.title = $(".post-title").text();
 }