+++ packages/telescope-tags/lib/client/routes.js
@@ -1,7 +1,5 @@
-preloadSubscriptions = typeof preloadSubscriptions === 'undefined' ? [] : preloadSubscriptions;
 preloadSubscriptions.push('categories');
 
-adminNav = typeof adminNav === 'undefined' ? [] : adminNav;
 adminNav.push({
   route: 'categories',
   label: 'Categories'