+++ server/migrations.js
@@ -1,66 +1,64 @@
+// TODO: switch over to Tom's migration package.
+
 // database migrations
 // http://stackoverflow.com/questions/10365496/meteor-how-to-perform-database-migrations
 Migrations = new Meteor.Collection('migrations');
 
 Meteor.startup(function () {
+  allMigrations = Object.keys(migrationsList);
+  _.each(allMigrations, function(migrationName){
+    runMigration(migrationName);
+  });
+});
 
+// wrapper function for all migrations
+var runMigration = function (migrationName) {
+  // migration updatePostStatus: make sure posts have a status
+  if (!Migrations.findOne({name: migrationName})) {
+    console.log("//----------------------------------------------------------------------//");
+    console.log("//------------//    Starting "+migrationName+" Migration    //-----------//");
+    console.log("//----------------------------------------------------------------------//");
+    Migrations.insert({name: migrationName, startedAt: new Date(), completed: false});
 
- // migration updatePostStatus: make sure posts have a status
-  if (!Migrations.findOne({name: "updatePostStatus"})) {
-    console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//    Starting updatePostStatus Migration    //-----------//");
-    console.log("//----------------------------------------------------------------------//");
-    Posts.find({status: {$exists : false}}).forEach(function (post) {
-        Posts.update(post._id, {$set: {status: 2}});
+    // execute migration function
+    var itemsAffected = migrationsList[migrationName]() || 0;
 
-        // START CONSOLE LOGS
-        console.log("---------------------");
-        console.log("Post: "+post.headline);
-        console.log("Updating status to approved");
-        // END CONSOLE LOGS
-      
-    });
-    Migrations.insert({name: "updatePostStatus"});
+    Migrations.update({name: migrationName}, {$set: {finishedAt: new Date(), completed: true, itemsAffected: itemsAffected}});
     console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//     Ending updatePostStatus Migration     //-----------//");
+    console.log("//------------//     Ending "+migrationName+" Migration     //-----------//");
     console.log("//----------------------------------------------------------------------//");
   }
+}
 
-
-
-
- // migration updateCategories: make sure categories have slugs
-  if (!Migrations.findOne({name: "updateCategories"})) {
-    console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//    Starting updateCategories Migration    //-----------//");
-    console.log("//----------------------------------------------------------------------//");
-    Categories.find().forEach(function (category) {
-      if(typeof category.slug === "undefined"){
+var migrationsList = {
+  updatePostStatus: function () {
+    var i = 0;
+    Posts.find({status: {$exists : false}}).forEach(function (post) {
+      i++;
+      Posts.update(post._id, {$set: {status: 2}});
+      console.log("---------------------");
+      console.log("Post: "+post.title);
+      console.log("Updating status to approved");  
+    });
+    return i;
+  },
+  updateCategories: function () {
+    var i = 0;
+    Categories.find({slug: {$exists : false}}).forEach(function (category) {
+        i++;
         var slug = slugify(category.name);
         Categories.update(category._id, {$set: {slug: slug}});
-
-        // START CONSOLE LOGS
         console.log("---------------------");
         console.log("Category: "+category.name);
         console.log("Updating category with new slug: "+slug);
-        // END CONSOLE LOGS
-      }
+    
     });
-    Migrations.insert({name: "updateCategories"});
-    console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//     Ending updateCategories Migration     //-----------//");
-    console.log("//----------------------------------------------------------------------//");
-  }
-
-
-
-
-  // migration updateCategories: store full category object in post instead of just the name
-  if (!Migrations.findOne({name: "updatePostCategories"})) {
-    console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//  Starting updatePostCategories Migration  //-----------//");
-    console.log("//----------------------------------------------------------------------//");
+    return i;
+  },
+  updatePostCategories: function () {
+    var i = 0;
     Posts.find().forEach(function (post) {
+      i++;
       var oldCategories = post.categories;
       var newCategories = [];
       var category = {};
@@ -89,7 +87,7 @@ Meteor.startup(function () {
 
       // START CONSOLE LOGS
       console.log("---------------------");
-      console.log("Post: "+post.headline);
+      console.log("Post: "+post.title);
       if(updating){
         console.log(oldCategories.length+" categories: "+oldCategories);
         console.log("Updating categories array to: ");
@@ -98,24 +96,16 @@ Meteor.startup(function () {
         console.log("No updates");
       }
       // END CONSOLE LOGS
-
     });
-    Migrations.insert({name: "updatePostCategories"});
-    console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//     Ending updateCategories Migration     //-----------//");
-    console.log("//----------------------------------------------------------------------//");
-  }
-
-  // migration updateUserProfiles: update user profiles with slugs and a few other properties
-  if (!Migrations.findOne({name: "updateUserProfiles"})) {
-    console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//  Starting updateUserProfiles Migration  //-----------//");
-    console.log("//----------------------------------------------------------------------//");
-
+    return i;
+  },
+  updateUserProfiles: function () {
+    var i = 0;
     var allUsers = Meteor.users.find();
     console.log('> Found '+allUsers.count()+' users.\n');
 
     allUsers.forEach(function(user){
+      i++;
       console.log('> Updating user '+user._id+' ('+user.username+')');
 
       // update user slug
@@ -135,10 +125,125 @@ Meteor.startup(function () {
       Meteor.users.update(user._id, {$set: {commentCount: commentsByUser.count()}});
 
     });
-    Migrations.insert({name: "updateUserProfiles"});
-    console.log("//----------------------------------------------------------------------//");
-    console.log("//------------//     Ending updateUserProfiles Migration     //-----------//");
-    console.log("//----------------------------------------------------------------------//");
+    return i;
+  },
+  resetUpvotesDownvotes: function () {
+    var i = 0;
+    Posts.find().forEach(function (post) {
+      i++;
+      var upvotes = 0,
+          downvotes = 0;
+      console.log("Post: "+post.title);
+      if(post.upvoters){
+        upvotes = post.upvoters.length;
+        console.log("Found "+upvotes+" upvotes.")
+      }
+      if(post.downvoters){
+        downvotes = post.downvoters.length;
+        console.log("Found "+downvotes+" downvotes.")
+      }
+      Posts.update(post._id, {$set: {upvotes: upvotes, downvotes: downvotes}});
+      console.log("---------------------");
+    });
+    return i;
+  },
+  resetCommentsUpvotesDownvotes: function () {
+    var i = 0;
+    Comments.find().forEach(function (comment) {
+      i++;
+      var upvotes = 0,
+          downvotes = 0;
+      console.log("Comment: "+comment._id);
+      if(comment.upvoters){
+        upvotes = comment.upvoters.length;
+        console.log("Found "+upvotes+" upvotes.")
+      }
+      if(comment.downvoters){
+        downvotes = comment.downvoters.length;
+        console.log("Found "+downvotes+" downvotes.")
+      }
+      Comments.update(comment._id, {$set: {upvotes: upvotes, downvotes: downvotes}});
+      console.log("---------------------");
+    });
+    return i;
+  },
+  headlineToTitle: function () {
+    var i = 0;
+    Posts.find({title: {$exists : false}}).forEach(function (post) {
+      i++;
+      console.log("Post: "+post.headline+" "+post.title);
+      Posts.update(post._id, { $rename: { 'headline': 'title'}}, {multi: true, validate: false});
+      console.log("---------------------");
+    });
+    return i;
+  },
+  commentsSubmittedToCreatedAt: function () {
+    var i = 0;
+    Comments.find().forEach(function (comment) {
+      i++;
+      console.log("Comment: "+comment._id);
+      Comments.update(comment._id, { $rename: { 'submitted': 'createdAt'}}, {multi: true, validate: false});
+      console.log("---------------------");
+    });
+    return i;
+  },
+  commentsPostToPostId: function () {
+    var i = 0;
+    Comments.find({postId: {$exists : false}}).forEach(function (comment) {
+      i++;
+      console.log("Comment: "+comment._id);
+      Comments.update(comment._id, { $rename: { 'post': 'postId'}}, {multi: true, validate: false});
+      console.log("---------------------");
+    });
+    return i;
+  },
+  createdAtSubmittedToDate: function () {
+    var i = 0;
+    Posts.find().forEach(function (post) {
+      if(typeof post.submitted == "number" || typeof post.createdAt == "number"){
+        i++;
+        console.log("Posts: "+post.title);
+        var createdAt = new Date(post.createdAt);
+        var submitted = new Date(post.submitted);
+        console.log(createdAt)
+        Posts.update(post._id, { $set: { 'createdAt': createdAt, submitted: submitted}}, {multi: true, validate: false});
+        console.log("---------------------");
+      }
+    });
+    return i;
+  },
+  commentsCreatedAtToDate: function () {
+    var i = 0;
+    Comments.find().forEach(function (comment) {
+      if(typeof comment.createdAt == "number"){
+        i++;
+        console.log("Comment: "+comment._id);
+        var createdAt = new Date(comment.createdAt);
+        console.log(createdAt)
+        Comments.update(comment._id, { $set: { 'createdAt': createdAt}}, {multi: true, validate: false});
+        console.log("---------------------");
+      }
+    });
+    return i;
+  },
+  submittedToPostedAt: function () {
+    var i = 0;
+    Posts.find({postedAt: {$exists : false}}).forEach(function (post) {
+      i++;
+      console.log("Post: "+post._id);
+      Posts.update(post._id, { $rename: { 'submitted': 'postedAt'}}, {multi: true, validate: false});
+      console.log("---------------------");
+    });
+    return i;
+  },
+  addPostedAtToComments: function () {
+    var i = 0;
+    Comments.find({postedAt: {$exists : false}}).forEach(function (comment) {
+      i++;
+      console.log("Comment: "+comment._id);
+      Comments.update(comment._id, { $set: { 'postedAt': comment.createdAt}}, {multi: true, validate: false});
+      console.log("---------------------");
+    });
+    return i;
   }
-
-});
+}